optimized app and added button to work with tests
